+++ src/lib/components/editor/BootstrapIconProvider.js
