+++ src/lib/components/editor/DCIconProvider.js
