+++ src/lib/components/editor/GridComponent.js
